Add beach_id to AttractionSerializer
This will make it much easier and less time-consuming to find the attractions belonging to a beach.
This allows me to match the beach_id of each attraction (from the "included" array) to a beach's id,
instead of iterating through the beach's attractions.data array and finding each corresponding attraction from the "included" array.